Moved form state to add project form
